+++ pkg/utils/common.go
@@ -1,0 +1,28 @@
+// Copyright (c) 2022 Intel Corporation.  All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License")
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package utils
+
+import (
+	"fmt"
+	"os"
+)
+
+func GetNodeIPFromEnv() (ipAddr string, err error) {
+	ipAddr = os.Getenv("NODE_IP")
+	if len(ipAddr) == 0 {
+		err = fmt.Errorf("NODE_IP env variable is not set")
+	}
+	return ipAddr, err
+}
